Revert "display-vnc: fix zoom-level set by command line"
This reverts commit 68148e1bd1a47ff370c78e2569a57ae0f3d8a400.

The commit in question was causing a regression about a tiny window
being displayed when opening a vnc guest.
The issue only happens with gtk-vnc/gtk2/virt-viewer versions present
on RHEL-6.7 and does not affect (gtk2 versions of) the upstream
packages, being this the reason to keep the patch downstream only.

Resolves: rhbz#1201679