virt-viewer-window: Set zoom when display is enabled and ready.
The display has default dimensions (100x100) when it is disabled.
Calculating the minimal zoom for the display will give wrong value
for the newly opened display.

It is better to wait for setting the zoom level to the moment when
the display is enabled and ready.

Related to: https://bugzilla.redhat.com/show_bug.cgi?id=1206460

(cherry picked from commit 998b13c881480ffea98dd8b65e0495c49fc31d2d)