virt-viewer-window: Return early when zoom of window and zoom of display are same
The zoom level should be changed when zoom levels of the window and
the display are different. It is wrong to check the previous value of
the window because it could be set just for the window and not for
the display (e.g. when setting zoom level using the command line).

Related to: https://bugzilla.redhat.com/show_bug.cgi?id=1206460

(cherry picked from commit 01249f703263ef6a8be59479df5cf6c5e8b1ceb9)