app: Set the resource base path
GApplication's resource base path is based on the application-id, for
instamce:
 - virt-viewer's resource base path: /org/virt-manager/virt-viewer, as
   the virt-viewer's application id is: org.virt-manager.virt-viewer.
 - remote-viewer's resource bash path: /org/virt-manager/remote-viewer
   as remote-viewer's application id is: org.virt-manager.remote-viewer

It's a issue because our resources have /org/virt-manager/virt-viewer
and Gtk, when trying to automatically load ui files (as done for
gtk/menu.ui, gtk/menus-appmenu.ui, gtk/menus-tradicional.ui and
gtk/help-overlay.ui), searches for these files in the base path.

For solving this issue, we can basically set the resource path using
g_application_set_resource_base_path() method. A check could be done and
this method could be called only when running remote-viewer, but as it's
a simple call, called only when the application starts I decided to go
without the application-id's check.

g_application_set_resource_base_path() was introduced in GLib 2.42 and
that's the reason I'm also bumping GLib dependency's version. Currently
it makes impossible to build virt-viewer on SLES 12 SP1 as it still has
GLib 2.38, so postponing this patch till SLES 12 SP2 release is
desirable.

Signed-off-by: Fabiano Fidêncio <fabiano@fidencio.org>