app: Use app-menu
Add support for an app-menu containing "Preferences", "About" and "Quit"
options.

The decision of naming the file as "menus.ui" under
"$(srcdir)/resources/gtk/" and having the widget id as "app-menu" is
that it can get loaded automatically during GApplication startup.

The options presented in the app-menu got removed from the
virt-viewer interface.

Signed-off-by: Sagar Ghuge <ghugesss@gmail.com>